use pnpm, add dockerfile (#11)
Buchung uses pnpm so it is only natural that nwd also uses pnpm

also checks the repo for spelling

also adds a dockerfile for hosting the nwd-api-server project.

also pushes docker image to github registry. It only pushes the latest
tag, when merged into main. This seems like a good starting point. After
pushing the docker image to the registry it is ready to be hosted
anywhere (render.com i our case).

In the future i guess a more elaborate release procedure (including zero
downtime, support for rollback, maybe staging, etc) could be useful. But
let's cross that bridge when we get there!

- fixes
https://github.com/buchungapp/nationaal-watersportdiploma/issues/10
- fixes
https://github.com/buchungapp/nationaal-watersportdiploma/issues/7