Don't throw an application error if a auth user has no personalities (#201)
Previously, an error occurred when a user tried to log in with an
unknown email address. Although the system would create an
authentication user, an error would be triggered because no personas
were attached. Now, we display a simple error message indicating that
the authentication user is not associated with any locations yet.
While add it, also improve some mobile UI buttons.