Location data in database (#329)
Since we are only using "useCache" and not "dynamicIO" the cacheLife
function is not applied. This caused the requests for location data to
be made every time.

To fix this the data is now stored in the metadata of a location under
"googlePlaceData" as enabling dynamicIO cost to much time for now

---------

Co-authored-by: Maurits Misana <maurits@techmaus.nl>