NWD-54 | fix: diploma download not working a second time after router.push, use window.open instead (#385)
### Problem
When downloading bulk diploma's for a second time: 
- The download never starts
- A POST request is made instead of a GET request, resulting in 405

The original code used the `redirect(redirectURL)` in the action, the
current code uses `router.push(redirectURL)`
Both have the same problem. The current solution is to perform a window
reload

### Findings
The POST request that causes the 405 is the server action that is called
the second time, but instead of posting to `/diplomas` or
`/cohort/.../diplomas` the post request is made to the redirected url.
Some where in Next the redirect url from the previous download is used
as the current window location.

### Solution
Instead of redirecting, or pushing on the same tab, open a new tab using
`window.open(redirectURL, "_blank")`